+++ views/partials/forms.settings.aside.php
@@ -1,0 +1,3 @@
+<div class="uk-margin">
+    <field-boolean bind="form.gui_in_header" label="@lang('Show in header menu')"></field-boolean>
+</div>