+++ views/partials/singletons.settings.aside.php
@@ -1,0 +1,3 @@
+<div class="uk-margin">
+    <field-boolean bind="singleton.gui_in_header" title="@lang('Show in header menu')" label="@lang('Show in header menu')"></field-boolean>
+</div>