+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* changed the key for custom navigation in header from `in_menu` to `gui_in_header` and added custom settings (Cockpit dropped that feature in version 0.9.3)
+
 ## 0.1.5
 
 * added option for custom navigation to settings, disabled by default